LoadSubTreeTask to dedicated file. fix namespaces
