+++ Assets/GeoTileLoader/Scripts/LoadSubTreeTask.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ee2225745edee4b16a7f6779c83f4322
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
